Move v1 applications api to api/v1 folder
