+++ internal/gateway/api/v1/application/handler/mockgatewayapplicationservice.go
