+++ internal/gateway/api/v1/application/service/mocksparkapplicationrepository.go
